Add auth check

+++ graphql/resolvers/booking.resolver.js
@@ -6,7 +6,12 @@ const Event = require('../../models/event');
 
 
 module.exports = {    
-    bookings: async () => {
+    bookings: async (args, req) => {
+
+        if(!req.isAuth){
+            throw new Error('Not authenticated');
+        }
+
         try{
             const bookings = await Booking.find();
             return bookings.map(booking => {
@@ -19,24 +24,35 @@ module.exports = {
     },
       
 
-    bookEvent: async args => {
+    bookEvent: async (args, req) => {
+
+        if(!req.isAuth){
+            throw new Error('Not authenticated');
+        }
+
         const fetchedEvent = await Event.findOne({ _id: args.eventId });
         const booking = new Booking({
-          user: '61b7da5a56b89b783addfb1b',
+        //   user: '61b7da5a56b89b783addfb1b',
+        user: req.userId,
           event: fetchedEvent
         });
         const result = await booking.save();
         return transformBooking(result);
       },
 
-    cancelBooking: async args => {
-    try {
-        const booking = await Booking.findById(args.bookingId).populate('event');
-        const event = transformEvent(booking.event);
-        await Booking.deleteOne({ _id: args.bookingId });
-        return event;
-    } catch (err) {
-        throw err;
-    }
+    cancelBooking: async (args, req) => {
+
+        if(!req.isAuth){
+            throw new Error('Not authenticated');
+        }
+
+        try {
+            const booking = await Booking.findById(args.bookingId).populate('event');
+            const event = transformEvent(booking.event);
+            await Booking.deleteOne({ _id: args.bookingId });
+            return event;
+        } catch (err) {
+            throw err;
+        }
     }
 }
